requirements.txt
# Sahayak_09 requirements

# Frontend
streamlit

# Backend & API
fastapi
uvicorn
pydantic
requests
pdfplumber
python-multipart
Pillow
# Optional: Only one vector DB, not both
chromadb
# OR
# cosdata-sdk @ git+https://github.com/cosdata/cosdata-sdk-python.git
